add weapon damtypes + other stuff

+++ src/player-util.h
@@ -62,7 +62,7 @@ enum
 
 struct monster_race *lookup_player_monster(const struct player *p);
 void check_player_monster(struct player *p, bool init);
-void player_race_name(struct player *p, char *buf, int bufsize);
+void player_race_name(struct player *p, char *buf, size_t bufsize);
 bool player_increase_stat(struct player *p);
 int get_power_scale(struct player *p, int power);
 int dungeon_get_next_level(struct player *p, int dlev, int added);
